requirements-macos.txt
# RapidWhisper Requirements - macOS
# Все зависимости для macOS (без pywin32, с pyobjc)

altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
colorama==0.4.6
coverage==7.13.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.151.4
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
keyboard==0.13.5
numpy==2.4.1
openai==2.16.0
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
psutil==7.2.2
pyaudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyperclip==1.11.0
pyqt6==6.10.2
pyqt6-qt6==6.10.2
pyqt6-sip==13.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-qt==4.5.0
python-dotenv==1.2.1
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
setuptools==80.10.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
